Add RUN_IN_CLUSTER_SETUP option to local-setup
local-setup is intended for local develoipment, and mostly you would be
running the controller locally. Currently this is not possible becasue
it's assuming the controller will be running in a kind cluster with
internal access (--internal option used when extracting the kubeconfig).

The RUN_IN_CLUSTER_SETUP option allows you do either, but defaults to
false to allow local development to work as normal.

To run local-setup when you intend to run the controller on the cluster,
use:
```
RUN_IN_CLUSTER_SETUP=true make local-setup
```

Also removes externalDNS from local setup again.