Fix the controller image build in GH actions
The Dockerfile now contains build target for both the controller and the
agent, so we need to specify which target we want to build to the build action
(right now is building just the last one, which happens to be the
agent). 'redhat-actions/buildah-build@v2' doesn't have support for the
'target' paremeter, so I have switched to the
'docker/build-push-action@v4' action instead.