Generate kustomize configs per cluster during local setup
Update the local-setup logic to generate a set of kustomize configs for
each kind cluster (control, or workload) being created during local
setup. These configs use one of two different overlays (control-cluster
or workload-cluster) depending on what type of cluster is being created,
and this overlay is responsible for maintaining the set of resources
that need to be added to that type of cluster.  Kustomize now manages
the creation of each clusters secret used by both argocd and the mctc
controller/agent/syncer itself. During generation of the clusters config
a `cluster-secret` kustomization is generated containing config
extracted from the running clusters kubeconfig. Note the `server` for
these secrets now use the internal node ips instead for the mctc
network, so there is no need for 2 secrets (internal/external).

In some cases during initial installation it is necassery to install
some resources or services before adding others i.e. nginx, base/istio
etc.. In order to allow this each overlay has an init kustomization that
contains a list of resources that must be installaed ahead of everything
else. After initial creation the default clusters kustomization can be
used to reconcile all changes as normal.