Enable RawFeedbackJsonString feature gate in OCM
Pull back the full array of Gateway addresses as json,
then unmarshall them into GatewayAddress objects.

This change requires an updated clusteradm cli (> 0.5.1)
and an updated version of ocm (>= 0.11.0)

Note there is an issue with the 0.11.0 ocm placement image on m1 macs.
See https://kubernetes.slack.com/archives/C01GE7YSUUF/p1685016272443249
for discussion.
The workaround in the commit is to patch the clustermanager to use
an amd64 build.