Geo DNS Strategy
Adds a LoadBalancing section to the DNSPolcy spec allowing default
weight and Geo code to be set.

LoadBalancing.Weighted.Default is required (default 120), and if set
will be used as the weight for a target clusters ednpoints if no other
weight can be determined from cluster attributes.
LoadBalancing.Geo.Default is optional, and when set indiciates the
default geo to use for target cluster endpoints if no other geo can be
determined from cluster attributes.

The generated DNSRecord structure has been updated to accomodate Geo and
Weighted only strategies using a combination of CNAMES and A records.