Fix gateway status updates
Fixes an issue with gateway status updates where it could end up in a
loop of constantly updating. This can be seen when placing gateways on
more than one cluster.
Orders the cluster list so that the status condition update always
prints the same message for the clusters the gateway is placed on.
Use the existing accepted condition if it already exists, currently
causes an update on every reconcile since `lastTransaction` is going to
be different on a new condition.
Replaces most of the condition helpers with the meta pkg equivalent
helpers.