Add prometheus and thanos for metrics federation.
Adds an optional env var, METRICS_FEDERATION, defaults to empty.
If set during local-setup, thanos is installed in the hub cluster
and prometheus in the spoke clusters.
The prometheus instance in the spoke clusters is configured to
remote write metrics to the thanos receive instance in the hub.

A thanos query instance is available via an Ingress in the hub
at https://thanos-query.172.32.0.2.nip.io