Fix dnsrecord when no addresses in any targets
Fixes an issue where a DNSRecord could get created with some endpoints,
but have no actual targets for them to point to.

If we have cluster targets that have no addresses when building up our
list of Endpoints for the DNSRecord they are now completely ignored and
no CNAME or A records related to that cluster are created.  If no
targets in a geo have valid addresses, the geo is ignored and no records
created for it.  If no records can be created for any targets (no
targets have addresses), nothing is created and an empty endpoints list
returned.