Handle MangedCluster delete events
When a ManagedCluster is deleted, don't attempt to update the DNSPolicy
straight away.
Instead, reconcile the Gateway first, which will trigger a DNSPolicy
reconcile if there are any changes related to placement changes e.g.
IPAddresses changing.

When a ManagedCluster is deleted, don't wait for the grace period
before deleting a Gateway (via ManifestWork).
Instead, delete the ManifestWork straight away.
Due to a known issue in OCM, https://github.com/open-cluster-management-io/ocm/issues/140
the ManifestWork is not removed after the deletionTimestamp is added
in this sceanrio.
To allow for this, the gateway placement logic will filter out any
ManifestWork resources that have a deletionTimestamp set.