DNSPolicy geo and weight validation updates
Removes the validation on the geo and weight fields in the DNSPolicy
spec in order to more easily support providers with more limited APIs.
The values for these fields are now provider specific, and as such any
invalid values will result in a failure in the API request resulting in
a DNSRecord reconcile error. The Route53 API requires the setting of
different fields in the API request depending on whether it's an ISO3166
Alpha-2 country or continent code, as such the validation of the country
code is now done in the AWS provider itself, and anything that isn't a
valid Alpha-2 code is assumed to be a continent.

Make the loadBalancing field in the DNSPolicy optional. If it's not set,
the default weight of 120 will be applied to all weighted records.