Add verify bundle
Adds a new verify bundle command that executes `make bundle` and ensures
the contents of `./bundle` are unchanged.

Add a check for createdAt being the only change to the bundle and ignore
it. Since operator-sdk 1.26.0, `make bundle` changes the `createdAt`
field from the bundle even if there are not other changes.