Remove cached helm charts during cleanup
Removes all cached helm chart directories from config during cleanup
(make clean or make local-cleanup).  There is a bug in kustomize where
the helm chart chached locally is never updated when the version of the
chart in the kustomization is changed. This could potentially lead to
different versions being used locally in development and running the
quick start.