Fix gateway controller deploy target
`make deploy-gateway-controller` deploys both the gateway and policy
controller but was only updating the image for the gateway controller to
the locally built version. This chnage updates the make targets to
ensure that both the gateway and policy controller images are updated
when testing the combined gateway/policy controller installation (e2e
tests).